Area of Circle, and Right Triangle
This program take a point you choose and creates a circle and find the area. It then uses the same point to create a Right Triangle and finds the area of it. It also finds the distance between the origin and the circle you choose and also the distance between two circles.