PHP/HTML/SQL
I was unable to get all the files from this class, but these are some I still had saved. The final project was a to create a social media site using PHP, HTML, and SQL to store users and their posts and comments. In order to be allowed to create a post you would need to log in first and if you were not a registered user it would give you an error message.